# docs/swagger.yaml
definitions:
...
  models.ErrorResponse:
    properties:
      code:
        example: 200
        type: integer
      error:
        example: Error message...
        type: string
      is_success:
        example: true
        type: boolean
      message:
        example: Example message success...
        type: string
      page:
        example: 1
        type: integer
    type: object
...
  models.ResponseData:
    properties:
      code:
        example: 200
        type: integer
      data: {}
      is_success:
        example: true
        type: boolean
      message:
        example: Example message success...
        type: string
      page:
        example: 1
        type: integer
    type: object
...
  models.TopUpRequest:
    properties:
      amount:
        type: integer
      payment_id:
        type: integer
      tax:
        type: integer
    required:
    - amount
    - payment_id
    type: object
...
host: 127.0.0.1:2409
info:
  contact: {}
  description: E-wallet team belalai
  title: Belalai E-Wallet
  version: "1.0"
paths:
...
  /topup:
    post:
      consumes:
      - application/json
      description: Create a new topup transaction with selected payment method
      parameters:
      - description: Topup request payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.TopUpRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.ResponseData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - JWTtoken: []
      summary: Create topup transaction
      tags:
      - TopUp
  /topup/methods:
    get:
      consumes:
      - application/json
      description: Retrieve list of supported payment methods for topup
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseData'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - JWTtoken: []
      summary: Get available payment methods
      tags:
      - TopUp
...
securityDefinitions:
  JWTtoken:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"